big clean up pass

+++ kbEngine/renderer/render_defs.cpp
@@ -1,19 +1,20 @@
-/// render_buffer.cpp
+/// render_defs.cpp
 ///
 /// 2025 kbEngine 2.0
 
-#include <Windows.h>
 #include <vector>
 #include "kbCore.h"
 #include "kbRenderer_defs.h"
-#include "render_buffer.h"
+#include "render_defs.h"
 
+/// RenderBuffer::write_vertex_buffer
 void RenderBuffer::write_vertex_buffer(const std::vector<vertexLayout>& vertices) {
-	m_num_elements = vertices.size();
+	m_num_elements = (uint32_t)vertices.size();
 	m_size_bytes = m_num_elements * sizeof(vertexLayout);
 }
 
+/// RenderBuffer::write_index_buffer
 void RenderBuffer::write_index_buffer(const std::vector<uint16_t>& indices) {
-	m_num_elements = indices.size();
+	m_num_elements = (uint32_t)indices.size();
 	m_size_bytes = m_num_elements * sizeof(uint16_t);
 }
